+++ irf/gadf/__init__.py
@@ -1,0 +1,9 @@
+from . import time
+from . import response
+from . import hdus
+
+__all__ = [
+    'time',
+    'response',
+    'hdus'
+]
